sfdp: modify tenant_index type from u16 to u32
* 'Tenant_index' was previously set on 16 bits. This
PR modify this so that it is set on 32 bits and allow
for more tenants to be specified.

* Add typedef 'sfdp_tenant_index_t' and 'sfdp_tenant_id_t'.

* Fix sections which relied on tenant_index being set on
16 bits e.g. invalid tenant_index value was previously considered
as ((u16) ~0) in certain sfdp services, and has been updated to
be (u32) (~0).

* Fix formatting issue to print tenant_id as an unsigned value.

Type: improvement

+++ src/vnet/sfdp/lookup/icmp_error_node.c
@@ -140,7 +140,7 @@ sfdp_lookup_icmp_inline (vlib_main_t *vm, vlib_node_runtime_t *node,
       while (n_left)
 	{
 	  uword flow_thread_index;
-	  u16 tenant_index;
+	  u32 tenant_index;
 	  sfdp_tenant_t *tenant;
 	  clib_bihash_kv_24_8_t kv4;
 
@@ -222,7 +222,7 @@ sfdp_lookup_icmp_inline (vlib_main_t *vm, vlib_node_runtime_t *node,
       while (n_left)
 	{
 	  uword flow_thread_index;
-	  u16 tenant_index;
+	  u32 tenant_index;
 	  sfdp_tenant_t *tenant;
 	  clib_bihash_kv_48_8_t kv6;
